+++ changelog.md
@@ -1,14 +1,22 @@
 # Changelog
 
+Latest version 2.5.11
+
 ## Unreleased
 
+* [Added] Orientation codes support
 * [Added] Data can be accessed by the property (e.g. `datap.data3d`)
+
+## 2.5
+
 * [Added] `read_dataset()` allow to read different datasets in the same way
 * [Changed] `join_path` return `Path` object as default
 * [Added] Datasets are now stored in `datasets.csv` on github
 * [Fixed] Deprecation Warning on `get_slice_location()`
+
+## 2.4
+
 * [Fixed] Shortest specific path now works
-* [Added] Orientation codes support
 
 ## 2.3